requirements.txt
PyQt5
Pillow
opencv-python
torch
torchvision
torchaudio
transformers
ftfy
regex
tqdm
spotipy
SQLAlchemy
